object: move implementation to encryptionkey.go
